Moved Test classes to matching package
