+++ src/test/java/main/resources/PostTweetResourceTest.java
@@ -1,4 +1,4 @@
-package main;
+package main.resources;
 
 import main.resources.PostTweetResource;
